backend/requirements.txt
fastapi>=0.101.0,<0.110.0
uvicorn==0.22.0
langchain-core==0.3.68
langgraph==0.3.5
langchain-groq==0.3.5
python-dotenv==1.0.0
requests==2.31.0